Mod: /domain/member 디렉토리 정리 - Member class, Guest class로 구분
