--- src/main/java/com/swyp/doubleSeven/domain/member/controller/TempController.java
@@ -1,4 +1,0 @@
-package com.swyp.doubleSeven.domain.member.controller;
-
-public class TempController {
-}
