--- src/main/java/com/swyp/doubleSeven/domain/member/dao/TempDao.java
@@ -1,4 +1,0 @@
-package com.swyp.doubleSeven.domain.member.dao;
-
-public class TempDao {
-}
