--- src/main/java/com/swyp/doubleSeven/domain/member/dto/TempDto.java
@@ -1,4 +1,0 @@
-package com.swyp.doubleSeven.domain.member.dto;
-
-public class TempDto {
-}
