+++ src/main/java/com/swyp/doubleSeven/domain/member/dto/response/KakaoTokenResponse.java
@@ -1,0 +1,10 @@
+package com.swyp.doubleSeven.domain.member.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+@Data
+public class KakaoTokenResponse {
+    @JsonProperty("access_token")
+    private String accessToken;
+}
