+++ src/main/java/com/swyp/doubleSeven/domain/member/dto/response/KakaoUserDTO.java
@@ -1,0 +1,12 @@
+package com.swyp.doubleSeven.domain.member.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class KakaoUserDTO {
+    private Long keyId;
+    private String nickname;
+    private String email;
+}
