+++ src/main/java/com/swyp/doubleSeven/domain/member/dto/response/MemberResponse.java
@@ -1,0 +1,23 @@
+package com.swyp.doubleSeven.domain.member.dto.response;
+
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Getter
+public class MemberResponse {
+    private Long memberId;
+    private String memberKeyId;
+    private String memberNickname;
+    private String memberName;
+    private String email;
+    private String role;
+    private String loginType;
+    private String dltnYn;
+    private LocalDateTime rgstDt;
+    private LocalDateTime updtDt;
+}
