--- src/main/java/com/swyp/doubleSeven/domain/member/kakao/dto/response/KakaoMemberResponse.java
@@ -1,19 +1,0 @@
-package com.swyp.doubleSeven.domain.member.kakao.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-import java.util.Map;
-
-@Data
-public class KakaoMemberResponse {
-    private Long id;
-    private Map<String, String> properties;
-    @JsonProperty("kakao_account")
-    private KakaoAccount kakaoAccount;
-
-    @Data
-    public static class KakaoAccount {
-        private String email;
-    }
-}
