--- src/main/java/com/swyp/doubleSeven/domain/member/kakao/dto/response/KakaoTokenResponse.java
@@ -1,10 +1,0 @@
-package com.swyp.doubleSeven.domain.member.kakao.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-@Data
-public class KakaoTokenResponse {
-    @JsonProperty("access_token")
-    private String accessToken;
-}
