--- src/main/java/com/swyp/doubleSeven/domain/member/kakao/dto/response/KakaoUserDTO.java
@@ -1,12 +1,0 @@
-package com.swyp.doubleSeven.domain.member.kakao.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class KakaoUserDTO {
-    private Long keyId;
-    private String nickname;
-    private String email;
-}
