--- src/main/java/com/swyp/doubleSeven/domain/member/login/LoginForm.java
@@ -1,15 +1,0 @@
-package com.swyp.doubleSeven.domain.member.login;
-
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Data;
-
-@Data
-public class LoginForm {
-
-    @NotEmpty
-    private String username;
-
-    @NotEmpty
-    private String password;
-
-}
