--- src/main/java/com/swyp/doubleSeven/domain/member/login/LoginService.java
@@ -1,9 +1,0 @@
-package com.swyp.doubleSeven.domain.member.login;
-
-import com.swyp.doubleSeven.domain.member.Member;
-
-public interface LoginService {
-
-    Member login(Member member);
-
-}
