+++ src/main/java/com/swyp/doubleSeven/domain/member/service/MemberService.java
@@ -1,24 +1,63 @@
 package com.swyp.doubleSeven.domain.member.service;
 
-import com.swyp.doubleSeven.domain.member.Member;
-import org.apache.ibatis.annotations.Mapper;
+import com.swyp.doubleSeven.domain.common.enums.Role;
+import com.swyp.doubleSeven.domain.member.dao.MemberDAO;
+import com.swyp.doubleSeven.domain.member.dto.reqeust.MemberRequest;
+import com.swyp.doubleSeven.domain.member.dto.response.KakaoUserDTO;
+import com.swyp.doubleSeven.domain.member.dto.response.MemberResponse;
+import com.swyp.doubleSeven.domain.common.enums.LoginType;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-@Mapper
-public interface MemberService {
+import java.time.LocalDateTime;
 
-    Member selectMemberByEmail(String email);
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+    private final KakaoApiClient kakaoApiClient;
+    private final MemberDAO memberDAO;
 
-    void insertMember(Member member);
+    public String getKakaoAccessToken(String authorizationCode) {
+        return kakaoApiClient.requestAccessToken(authorizationCode);
+    }
 
-    void updateMember(Member member);
+    public MemberResponse processKakaoUser(String accessToken) {
+        KakaoUserDTO kakaoUser = kakaoApiClient.getUserInfo(accessToken);
 
-    Member findByUsername(String username);
+        MemberResponse existingMember = memberDAO.findMemberByOauthProviderAndMemberId("KAKAO", kakaoUser.getKeyId());
 
-    Member findByEmail(String email);
+        MemberRequest.MemberRequestBuilder memberRequestBuilder = MemberRequest.builder()
+                .memberKeyId(String.valueOf(kakaoUser.getKeyId()))
+                .loginType(LoginType.KAKAO.getType())
+                .memberNickname(kakaoUser.getNickname())
+                .email(kakaoUser.getEmail())
+                .role(Role.MEMBER.getType())
+                .dltnYn("N")
+                .rgstId(0L)
+                .rgstDt(LocalDateTime.now())
+                .updtId(0L)
+                .updtDt(LocalDateTime.now());
 
-    Member findByPhone(String phone);
+        if(existingMember == null) {
+            // 신규 회원 처리
+            memberDAO.insertMember(memberRequestBuilder.build());
+        } else {
+            // 기존 회원 처리
+            memberRequestBuilder.memberId(existingMember.getMemberId())
+                    .updtId("GUEST".equals(existingMember.getLoginType()) ? 0L : existingMember.getMemberId())
+                    .memberNickname(existingMember.getMemberNickname())
+                    .email(existingMember.getEmail());
 
+            memberDAO.updateMember(memberRequestBuilder.build());
+        }
 
-
-
+        // MemberResponse 생성
+        return MemberResponse.builder()
+                .memberId(memberRequestBuilder.memberKeyId(String.valueOf(kakaoUser.getKeyId())).build().getMemberId())
+                .memberNickname(memberRequestBuilder.build().getMemberNickname())
+                .email(memberRequestBuilder.build().getEmail())
+                .role(memberRequestBuilder.build().getRole())
+                .loginType(memberRequestBuilder.build().getLoginType())
+                .build();
+    }
 }
