+++ src/main/resources/application.yml
@@ -4,10 +4,6 @@ server:
   servlet:
     context-path: /
 
-#
-
-
-
 # DataBase
 spring:
   datasource:
