--- src/main/resources/mybatis/mapper/member/member.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com/swyp/doubleSeven/domain/member/service/MemberService">
-
-    <!-- 이메일로 회원 조회 -->
-    <select id="selectMemberByEmail" parameterType="String" resultType="com.swyp.doubleSeven.domain.member.Member">
-        SELECT MEMBER_ID,
-               MEMBER_NICKNAME,
-               MEMBER_NAME,
-               EMAIL,
-               ROLE,
-               OAUTH_PROVIDER,
-               DLTN_YN,
-               RGST_ID,
-               RGST_DT,
-               UPDT_ID,
-               UPDT_DT
-        FROM MEMBER
-        WHERE EMAIL = #{email}
-          AND DLTN_YN = 'N'
-    </select>
-
-    <!-- 회원 삽입 -->
-    <insert id="insertMember" parameterType="com.swyp.doubleSeven.domain.member.Member">
-        INSERT INTO MEMBER (MEMBER_NICKNAME,
-                            MEMBER_NAME,
-                            EMAIL,
-                            ROLE,
-                            OAUTH_PROVIDER,
-                            DLTN_YN,
-                            RGST_ID,
-                            RGST_DT,
-                            UPDT_ID,
-                            UPDT_DT)
-        VALUES (#{memberNickname},
-                #{memberName},
-                #{email},
-                #{role},
-                #{oauthProvider},
-                #{dltnYn},
-                #{rgstId},
-                NOW(),
-                #{updtId},
-                NOW())
-    </insert>
-
-    <!-- 회원 업데이트 -->
-    <update id="updateMember" parameterType="com.swyp.doubleSeven.domain.member.Member">
-        UPDATE MEMBER
-        SET MEMBER_NAME = #{memberName},
-            UPDT_ID     = #{updtId},
-            UPDT_DT     = NOW()
-        WHERE MEMBER_ID = #{memberId}
-    </update>
-
-</mapper>
